server/requirements.txt
Django==3.0.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
djongo==1.3.4
Pillow==8.2.0
pymongo==3.11.3
PyJWT==1.7.1
django-allauth==0.44.0
django-mail-templated==2.6.5